- Removed Help window - Added link to Github page since there is more info there. - Doc strings on photo_viewer.py - added threaded_spread_count.py to host threading modules - added logic to compare version tags into core.py - Added logic to open a popup if user doesnt have the latest build. - Docs on project.py - added requests to requirements.txt
